Move ingestStartedStatus to IngestJobStatusFromJobTestData
